types test in controllers (#3)
