+++ test/types_test.go
@@ -1,0 +1,13 @@
+package test
+
+// test flash message: struct
+type flash struct {
+	success_message string
+	error_message   string
+}
+
+// test regex: struct
+type regex struct {
+	must_compile string
+	actual       string
+}
